+++ src/PluginsManager/Microsoft.Performance.Toolkit.Plugins.Runtime/Exceptions/PluginManagerException.cs
@@ -1,0 +1,28 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System;
+
+namespace Microsoft.Performance.Toolkit.Plugins.Runtime.Exceptions
+{
+    /// <summary>
+    ///    Base exception type for the PluginsManager.
+    /// </summary>
+    public abstract class PluginsManagerException
+       : Exception
+    {
+        protected PluginsManagerException()
+        {
+        }
+
+        protected PluginsManagerException(string message)
+            : base(message)
+        {
+        }
+
+        protected PluginsManagerException(string message, Exception innerException)
+            : base(message, innerException)
+        {
+        }
+    }
+}
